Minimize upstream dependencies
Since odl-bier-adapter is pulling in odl-netconf-console, which pulls
in odl-mdsal-broker, there is no need to pull that explicitly.

Since odl-bier-std-driver and others are pulling in odl-bier-adapter,
they do not need to pull odl-mdsal-broker nor odl-netconf-console
themselves. The same applies to odl-bier-bierman, odl-bier-service.

This cleans up the features, speeding up the build time and startup
time.

Change-Id: Ief4e0a17d2cf0955e5c0c4dac9705b623b010de3
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>